--- Script/limparModal.js
@@ -1,5 +1,0 @@
-$('#loginModal').on('hidden.bs.modal', function (e) {
-    // limpa o formulário quando o modal for fechado
-    $('form').trigger('reset');
-  });
-
