--- css/style.css
@@ -1,136 +1,0 @@
-/* Importa a fonte Muli do Google Fonts com a variação "display" */
-@import url('https://fonts.googleapis.com/css?family=Muli&display=swap');
-
-/* Define as propriedades básicas para todos os elementos do HTML */
-* {
-  margin: 0;
-  padding: 0;
-  border: 0;
-  box-sizing: border-box;
-}
-
-/* Define o fundo e a cor de texto do body para branco e preto, respectivamente */
-body {
-  background-color: white;
-  color: black;
-}
-
-/* Define o fundo e a cor de texto do body quando a classe "dark" é aplicada */
-body.dark {
-  background-color: black;
-  color: white;
-}
-
-main {
-  position: relative;
-  min-height: calc(100vh - 100px); /* altura mínima para ocupar a página inteira, descontando a altura do footer */
-  top: 50px; /* altura do elemento <head> */
-}
-
-
-header, footer {
-  position: fixed;
-  left: 0;
-  right: 0;
-  z-index: 100;
-}
-
-/* Esconde a checkbox e define sua posição como absoluta */
-.checkbox {
-  opacity: 0;
-  position: absolute;
-}
-
-/* Define um estilo de formatação para a caixa de seleção e seu respectivo rótulo */
-.label {
-  background-color: #111;
-  border-radius: 50px;
-  cursor: pointer;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 5px;
-  position: relative;
-  height: 26px;
-  width: 50px;
-  transform: scale(1.5);
-}
-
-/* Define um estilo de formatação para a bolinha dentro da caixa de seleção */
-.label .ball {
-  background-color: #fff;
-  border-radius: 50%;
-  position: absolute;
-  top: 2px;
-  left: 2px;
-  height: 22px;
-  width: 22px;
-  transform: translateX(0px);
-  transition: transform 0.2s linear;
-}
-
-/* Define um estilo de formatação para a bolinha dentro da caixa de seleção quando a checkbox está marcada */
-.checkbox:checked + .label .ball {
-  transform: translateX(24px);
-}
-
-/* Define a cor do ícone da lua */
-.fa-moon {
-  color: #f1c40f;
-}
-
-/* Define a cor do ícone do sol */
-.fa-sun {
-  color: #f39c12;
-}
-
-/* Define a margem superior da navegação */
-.nav {
-  margin-top: 30px;
-}
-
-/* Define o tamanho máximo da imagem de perfil */
-.profile img {
-  max-width: 150px;
-}
-
-.button-group {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-
-/* Define o estilo do menu */
-.menu {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-/* Define o estilo dos itens do menu */
-.menu li {
-  list-style: none;
-  margin-right: 20px;
-}
-
-/* Posiciona a lupa à direita em telas grandes */
-@media (min-width: 992px) {
-  .search-icon {
-    float: right;
-    margin-left: 10px;
-  }
-}
-
-/* Posiciona a lupa à direita em telas pequenas */
-@media (max-width: 991px) {
-  .search-icon {
-    position: fixed;
-    float: right;
-    margin-left: 10px;
-    right: 10px;
-    margin: 0;
-  }
-}
-
-
-
