--- hobbies.html
@@ -1,198 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-br">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="description" content="Página de exemplo com um layout semelhante ao GitHub.">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Rainer Teixeira Page</title>
-    <link rel="shortcut icon" href="img/favicon.png" />
-    <link rel="stylesheet" href="css/bootstrap.css">
-    <link rel="stylesheet" href="css/style.css">
-</head>
-
-<body>
-    <header class="bg-dark text-white">
-        <nav class="navbar navbar-expand-lg navbar-dark container">
-            <a class="navbar-brand" href="#">Rainer Teixeira</a>
-            <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav"
-                aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-                <span class="navbar-toggler-icon"></span>
-            </button>
-            <div class="collapse navbar-collapse" id="navbarNav">
-                <ul class="navbar-nav mr-auto">
-                    <!-- Item "Projetos" -->
-                    <li class="nav-item">
-                        <a class="nav-link" href="projetos.html"><i class="fas fa-laptop-code fa-lg"></i> Projetos</a>
-                    </li>
-                    <!-- Item "Sobre mim" -->
-                    <li class="nav-item">
-                        <a class="nav-link" href="sobre.html"><i class="fas fa-user-tie fa-lg"></i> Sobre mim</a>
-                    </li>
-                    <!-- Item "Hobbies" -->
-                    <li class="nav-item">
-                        <a class="nav-link" href="hobbies.html"><i class="fas fa-gamepad fa-lg"></i> Hobbies</a>
-                    </li>
-                    <!-- Item "Contato" -->
-                    <li class="nav-item">
-                        <a class="nav-link" href="contato.html"><i class="fas fa-comments fa-lg"></i> Contato</a>
-                    </li>
-                    <!-- Item "Blog" -->
-                    <li class="nav-item">
-                        <a class="nav-link" href="blog.html"><i class="fas fa-blog fa-lg"></i> Blog</a>
-                    </li>
-                </ul>
-                <div class="d-flex align-items-center ml-auto button-group">
-
-                    <div class="form-check">
-                        <input type="checkbox" class="checkbox" id="chk" />
-                        <label class="label" for="chk">
-                            <i class="fas fa-moon"></i>
-                            <i class="fas fa-sun"></i>
-                            <div class="ball"></div>
-                        </label>
-                    </div>
-                </div>
-            </div>
-        </nav>
-    </header>
-    <main>
-        <div class="container">
-            <div class="row">
-
-                <section class="col-md-9">
-                    <h1>HOBBIES</h1>
-                    <p>Esta e uma area restrita</p>
-
-                </section>
-
-
-
-                <section class="col-md-9">
-                    <h1>HOBBIES</h1>
-                    <p>Esta e uma area restrita</p>
-
-                </section>
-
-
-                <section class="col-md-9">
-                    <h1>HOBBIES</h1>
-                    <p>Esta e uma area restrita</p>
-
-                </section>
-
-
-                <section class="col-md-9">
-                    <h1>HOBBIES</h1>
-                    <p>Esta e uma area restrita</p>
-
-                </section>
-
-
-                <section class="col-md-9">
-                    <h1>HOBBIES</h1>
-                    <p>Esta e uma area restrita</p>
-
-                </section>
-
-
-                <section class="col-md-9">
-                    <h1>HOBBIES</h1>
-                    <p>Esta e uma area restrita</p>
-
-                </section>
-
-
-                <section class="col-md-9">
-                    <h1>HOBBIES</h1>
-                    <p>Esta e uma area restrita</p>
-
-                </section>
-
-
-                <section class="col-md-9">
-                    <h1>HOBBIES</h1>
-                    <p>Esta e uma area restrita</p>
-
-                </section>
-
-
-                <section class="col-md-9">
-                    <h1>HOBBIES</h1>
-                    <p>Esta e uma area restrita</p>
-
-                </section>
-
-
-
-            </div>
-        </div>
-    </main>
-    <footer class="bg-dark text-white py-3">
-        <div class="container">
-            <div class="row">
-                <div class="col-md-6">
-                    <h3>Entre em contato</h3>
-                    <p>Email: raineroliveira94@hotmail.com</p>
-                    <p>Telefone: (24) 99913-7382</p>
-                </div>
-                <div class="col-md-6">
-                    <h3>Direitos autorais</h3>
-                    <p>&copy; 2023 Página estilo GitHub. Todos os direitos reservados..</p>
-                </div>
-            </div>
-        </div>
-    </footer>
-    <!-- Links dos arquivos JavaScript -->
-    <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js"></script>
-    <script src="https://code.jquery.com/jquery.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"></script>
-    <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
-    <script src="Script/dark-mode.js"></script>
-    <script src="Script/login.js"></script>
-    <script src="https://kit.fontawesome.com/998c60ef77.js" crossorigin="anonymous"></script>
-
-    <!-- Modal de login -->
-    <div class="modal fade" id="loginModal" tabindex="-1" role="dialog" aria-labelledby="loginModalLabel"
-        aria-hidden="true">
-        <div class="modal-dialog" role="document">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h5 class="modal-title" id="loginModalLabel">Login</h5>
-                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                        <span aria-hidden="true">&times;</span>
-                    </button>
-                </div>
-                <div class="modal-body">
-                    <form>
-                        <div class="form-group">
-                            <label for="inputEmail">Email</label>
-                            <input type="email" class="form-control" id="inputEmail" placeholder="Digite seu email">
-                        </div>
-                        <div class="form-group">
-                            <label for="inputPassword">Senha</label>
-                            <input type="password" class="form-control" id="inputPassword"
-                                placeholder="Digite sua senha">
-                        </div>
-                    </form>
-                </div>
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-secondary" data-dismiss="modal">Fechar</button>
-                    <button type="button" class="btn btn-secondary" data-toggle="modal" data-target="#loginModal"
-                        id="btn-entrar">
-                        Login
-                    </button>
-
-                    <button type="button" class="btn btn-secondary" data-toggle="modal" data-target="#loginModal" id="btn-entrar">
-                        Login
-                    </button>
-                    
-
-                </div>
-            </div>
-        </div>
-    </div>
-</body>
-
-</html>
