+++ templates/Script/limparModal.js
@@ -1,0 +1,5 @@
+$('#loginModal').on('hidden.bs.modal', function (e) {
+    // limpa o formulário quando o modal for fechado
+    $('form').trigger('reset');
+  });
+
