+++ templates/css/read-mode.css
@@ -1,0 +1,45 @@
+@media read-mode {
+    /* Estilos para modo de leitura */
+    
+    /* Remover elementos não essenciais */
+    nav, header, footer, aside {
+    display: none;
+    }
+    
+    /* Aumentar o tamanho da fonte */
+    body {
+    font-size: 1.2em;
+    }
+    
+    /* Definir largura máxima de imagens */
+    img {
+    max-width: 100%;
+    }
+    
+    /* Ajustar a cor do texto para facilitar a leitura */
+    body {
+    color: #444;
+    background-color: #fff;
+    }
+    
+    /* Adicionar espaço extra para melhorar a legibilidade */
+    p, ul, ol {
+    margin-bottom: 1.5em;
+    }
+    
+    /* Ajustar o espaçamento entre linhas */
+    body {
+    line-height: 1.5;
+    }
+    
+    /* Ajustar a largura do conteúdo */
+    main {
+    max-width: 800px;
+    margin: 0 auto;
+    }
+    
+    /* Esconder os elementos de navegação de página */
+    .page-nav {
+    display: none;
+    }
+    }
