+++ templates/footer.html
@@ -1,0 +1,19 @@
+    <footer class="bg-dark text-white py-3">
+        <div class="container">
+            <div class="row text-center">
+                <div class="col-md-6">
+                    <!-- Informações de contato -->
+                    <h3 class="text-white">Entre em contato</h3>
+                    <p><i class="fa fa-envelope"></i> <a href="mailto:raineroliveira94@hotmail.com" class="text-white"
+                            target="_blank">raineroliveira94@hotmail.com</a></p>
+                    <p class="text-white"><i class="fa fa-phone"></i> <a href="tel:+5524999137382"
+                            class="text-white">(24) 99913-7382</a></p>
+                </div>
+                <div class="col-md-6">
+                    <!-- Direitos autorais -->
+                    <h3 class="text-white">Direitos autorais</h3>
+                    <p class="text-white">&copy; 2023 Rainer Pages. Todos os direitos reservados.</p>
+                </div>
+            </div>
+        </div>
+    </footer>
