+++ templates/header.html
@@ -1,0 +1,60 @@
+
+  <header class="bg-dark text-white custom-navbar">
+    <nav class="navbar navbar-expand-lg navbar-dark container">
+      <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
+        <span class="navbar-toggler-icon"></span>
+      </button>
+      <div class="d-flex align-items-center justify-content-between w-100">
+        <div class="navbar-brand logo-mobile-center navbar-brand-custom">
+          <a>Rainer Teixeira</a>
+        </div>
+        <div class="d-flex align-items-center d-lg-none">
+          <a class="nav-link text-white mr-3" href="#" data-toggle="modal" data-target="#loginModal">
+            <i class="fas fa-user fa-lg"></i>
+          </a>
+          <div class="form-check ml-3">
+            <input type="checkbox" class="checkbox" id="chk" />
+            <label class="label" for="chk">
+              <span class="toggle-icon"><i class="fas fa-sun"></i></span>
+              <span class="toggle-icon"><i class="fas fa-moon"></i></span>
+              <div class="ball"></div>
+            </label>
+          </div>
+        </div>
+        <div class="collapse navbar-collapse justify-content-end" id="navbarNav">
+          <ul class="navbar-nav mx-auto">
+            <li class="nav-item custom-nav-item">
+              <a class="nav-link" href="projetos.html"><i class="fas fa-laptop-code fa-lg"></i> Projetos</a>
+            </li>
+            <li class="nav-item custom-nav-item">
+              <a class="nav-link" href="sobre.html"><i class="fas fa-user-tie fa-lg"></i> Sobre mim</a>
+            </li>
+            <li class="nav-item custom-nav-item">
+              <a class="nav-link" href="hobbies.html"><i class="fas fa-gamepad fa-lg"></i> Hobbies</a>
+            </li>
+            <li class="nav-item custom-nav-item">
+              <a class="nav-link" href="contato.html"><i class="fas fa-comments fa-lg"></i> Contato</a>
+            </li>
+            <li class="nav-item custom-nav-item">
+              <a class="nav-link" href="blog.html"><i class="fas fa-blog fa-lg"></i> Blog</a>
+            </li>
+            <li class="nav-item d-none d-lg-block custom-nav-item">
+              <a class="nav-link text-white mr-3" href="#" data-toggle="modal" data-target="#loginModal">
+                <i class="fas fa-user fa-lg"></i>
+              </a>
+            </li>
+            <li class="nav-item d-none d-lg-block custom-nav-item">
+              <div class="form-check ml-3">
+                <input type="checkbox" class="checkbox" id="chk" />
+                <label class="label" for="chk">
+                  <span class="toggle-icon"><i class="fas fa-sun"></i></span>
+                  <span class="toggle-icon"><i class="fas fa-moon"></i></span>
+                  <div class="ball"></div>
+                </label>
+              </div>
+            </li>
+          </ul>
+        </div>
+      </div>
+    </nav>
+  </header>
