+++ templates/index.html
@@ -1,0 +1,180 @@
+<!DOCTYPE html>
+<html lang="pt-br">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="description" content="Página de exemplo com um layout semelhante ao GitHub.">
+    <meta name="viewport"
+        content="width=device-width, initial-scale=1, minimum-scale=1, maximum-scale=1, viewport-fit=cover">
+    <title>Rainer Teixeira Page</title>
+    <link rel="shortcut icon" href="img/favicon.png" />
+    <link rel="stylesheet" href="css/bootstrap.css">
+    <link rel="stylesheet" href="css/style.css">
+    <meta name="apple-mobile-web-app-capable" content="yes">
+    <meta name="apple-mobile-web-app-status-bar-style" content="black-translucent">
+    <meta name="theme-color" content="#000000">
+    <link rel="stylesheet" href="css/read-mode.css" media="read-mode">
+</head>
+
+<body>
+    <div id="header-placeholder"></div>
+
+    <main>
+        <div class="container">
+            <div class="row">
+                <aside class="col-md-3">
+                    <!-- Perfil do usuário -->
+                    <div class="profile text-center">
+                        <img src="https://avatars.githubusercontent.com/u/66490370?s=400&u=362ec0da5fdc3ec1765b59ade1442010c189eac1&v=4"
+                            alt="Profile picture" class="img-fluid rounded-circle mb-3">
+                        <h2>Rainer Teixeira</h2>
+                        <p>raineroliveira94@hotmail.com</p>
+                    </div>
+                    <!-- Navegação -->
+                    <nav class="nav flex-column">
+                        <a class="nav-link active" href="#">Overview</a>
+                        <a class="nav-link" href="#">Repositórios</a>
+                        <a class="nav-link" href="#">Projetos</a>
+                        <a class="nav-link" href="#">Estrelas</a>
+                        <a class="nav-link" href="#">Seguidores</a>
+                        <a class="nav-link" href="#">Seguindo</a>
+                    </nav>
+                </aside>
+                <section class="col-md-9">
+                    <div class="row">
+                        <div class="col-md-12">
+                            <!-- Título e introdução -->
+                            <h1 class="text-center mb-5">Bem vindo a minha pagina..</h1>
+                            <p class="text-center">Esta é uma página de exemplo simples com um layout semelhante ao
+                                GitHub..</p>
+                        </div>
+                    </div>
+                    <div class="row">
+                        <div class="col-md-6">
+                            <!-- Previsão do tempo -->
+                            <section class="weather-widget mb-5">
+                                <h3>Previsão do tempo</h3>
+                                <div class="weather-info">
+                                    <!-- informações do tempo serão exibidas aqui -->
+                                </div>
+                            </section>
+                        </div>
+                        <div class="col-md-6">
+                            <!-- Outra seção de previsão do tempo -->
+                            <section class="weather-widget mb-5">
+                                <h3>Previsão do tempo</h3>
+                                <div class="weather-info">
+                                    <!-- informações do tempo serão exibidas aqui -->
+                                </div>
+                                <div class="weather-info">
+                                    <!-- informações do tempo serão exibidas aqui -->
+                                </div>
+                            </section>
+                        </div>
+                    </div>
+                </section>
+            </div>
+        </div>
+    </main>
+
+    <div id="footer-placeholder"></div>
+
+    <!-- Incluindo jQuery antes de qualquer script que dependa dele -->
+    <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
+    <script>
+        $(function () {
+            $("#header-placeholder").load("header.html");
+            $("#footer-placeholder").load("footer.html");
+        });
+    </script>
+
+    <!-- Incluindo Bootstrap -->
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
+
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js"></script>
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"></script>
+    <!-- Scripts personalizados -->
+    <script src="Script/dark-mode.js"></script>
+    <script src="Script/tempo.js"></script>
+    <script src="Script/login.js"></script>
+    <script src="Script/read-mode.js"></script>
+    <script src="Script/pesquisar.js"></script>
+
+    <script>
+        $(function () {
+            var footer = $('#footer-placeholder');
+            var body = $('body');
+            var windowHeight = $(window).height();
+            var bodyHeight = body.height();
+
+            if (bodyHeight < windowHeight) {
+                footer.css({
+                    position: 'absolute',
+                    bottom: '0',
+                    width: '100%'
+                });
+            } else {
+                footer.css({
+                    position: 'static'
+                });
+            }
+
+            $(window).scroll(function () {
+                if ($(window).scrollTop() + $(window).height() == $(document).height()) {
+                    footer.css({
+                        position: 'static'
+                    }).fadeIn();
+                } else {
+                    footer.fadeOut();
+                }
+            });
+        });
+
+    </script>
+
+    <!-- Ícones do Font Awesome -->
+    <script src="https://kit.fontawesome.com/998c60ef77.js" crossorigin="anonymous"></script>
+
+    <!-- Modal de login -->
+    <div class="modal fade" id="loginModal" tabindex="-1" role="dialog" aria-labelledby="loginModalLabel"
+        aria-hidden="true">
+        <div class="modal-dialog" role="document">
+            <div class="modal-content">
+                <!-- Cabeçalho do modal -->
+                <div class="modal-header">
+                    <h5 class="modal-title" id="loginModalLabel">Login</h5>
+                    <!-- Botão de fechar modal -->
+                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                        <span aria-hidden="true">&times;</span>
+                    </button>
+                </div>
+                <!-- Corpo do modal -->
+                <div class="modal-body">
+                    <form>
+                        <div class="form-group">
+                            <label for="inputEmail">Email</label>
+                            <input type="email" class="form-control" id="inputEmail" placeholder="Digite seu email">
+                        </div>
+                        <div class="form-group">
+                            <label for="inputPassword">Senha</label>
+                            <input type="password" class="form-control" id="inputPassword"
+                                placeholder="Digite sua senha">
+                        </div>
+                    </form>
+                </div>
+                <!-- Rodapé do modal -->
+                <div class="modal-footer">
+                    <!-- Botão de fechar modal -->
+                    <button type="button" class="btn btn-secondary" data-dismiss="modal">Fechar</button>
+                    <!-- Botão de enviar formulário de login -->
+                    <button type="button" class="btn btn-secondary" data-toggle="modal" data-target="#loginModal"
+                        id="entrar-btn">
+                        Entrar
+                    </button>
+                </div>
+            </div>
+        </div>
+
+</body>
+
+</html>
